[bphh-613] Refactor requirement template domain folder structure for better organization
