+++ app/frontend/components/domains/requirement-template/screens/edit-requirement-template-screen/controls-header.tsx
@@ -5,7 +5,7 @@ import React from "react"
 import { useFormContext } from "react-hook-form"
 import { useTranslation } from "react-i18next"
 import { useNavigate } from "react-router-dom"
-import { IRequirementTemplate } from "../../../../models/requirement-template"
+import { IRequirementTemplate } from "../../../../../models/requirement-template"
 import { IRequirementTemplateForm } from "./index"
 import { PublishScheduleModal } from "./publish-schedule-modal"
