+++ app/frontend/components/domains/requirement-template/screens/edit-requirement-template-screen/editable-builder-header.tsx
@@ -3,9 +3,9 @@ import * as R from "ramda"
 import React from "react"
 import { useFormContext } from "react-hook-form"
 import { useTranslation } from "react-i18next"
-import { IRequirementTemplate } from "../../../../models/requirement-template"
-import { ETemplateVersionStatus } from "../../../../types/enums"
-import { EditableInputWithControls } from "../../../shared/editable-input-with-controls"
+import { IRequirementTemplate } from "../../../../../models/requirement-template"
+import { ETemplateVersionStatus } from "../../../../../types/enums"
+import { EditableInputWithControls } from "../../../../shared/editable-input-with-controls"
 import { BuilderHeader } from "./builder-header"
 import { IRequirementTemplateForm } from "./index"
