+++ app/frontend/components/domains/requirement-template/screens/edit-requirement-template-screen/index.tsx
@@ -6,20 +6,20 @@ import { FormProvider, useFieldArray, useForm } from "react-hook-form"
 import { useTranslation } from "react-i18next"
 import { RemoveScroll } from "react-remove-scroll"
 import { v4 as uuidv4 } from "uuid"
-import { useRequirementTemplate } from "../../../../hooks/resources/use-requirement-template"
-import { IRequirementTemplate } from "../../../../models/requirement-template"
-import { ITemplateSectionBlockModel } from "../../../../models/template-section-block"
-import { useMst } from "../../../../setup/root"
+import { useRequirementTemplate } from "../../../../../hooks/resources/use-requirement-template"
+import { IRequirementTemplate } from "../../../../../models/requirement-template"
+import { ITemplateSectionBlockModel } from "../../../../../models/template-section-block"
+import { useMst } from "../../../../../setup/root"
 import {
   IRequirementTemplateSectionAttributes,
   IRequirementTemplateUpdateParams,
   ITemplateSectionBlockAttributes,
-} from "../../../../types/api-request"
-import { ErrorScreen } from "../../../shared/base/error-screen"
-import { LoadingScreen } from "../../../shared/base/loading-screen"
-import { BuilderFloatingButtons } from "../builder-floating-buttons"
-import { SectionsSidebar } from "../sections-sidebar"
-import { useSectionHighlight } from "../use-section-highlight"
+} from "../../../../../types/api-request"
+import { ErrorScreen } from "../../../../shared/base/error-screen"
+import { LoadingScreen } from "../../../../shared/base/loading-screen"
+import { BuilderFloatingButtons } from "../../builder-floating-buttons"
+import { SectionsSidebar } from "../../sections-sidebar"
+import { useSectionHighlight } from "../../use-section-highlight"
 import { ControlsHeader } from "./controls-header"
 import { EditableBuilderHeader } from "./editable-builder-header"
 import { SectionsDisplay } from "./sections-display"
