+++ app/frontend/components/domains/requirement-template/screens/edit-requirement-template-screen/sections-display.tsx
@@ -6,12 +6,12 @@ import React from "react"
 import { useFieldArray, useFormContext } from "react-hook-form"
 import { useTranslation } from "react-i18next"
 import { v4 as uuidv4 } from "uuid"
-import { useMst } from "../../../../setup/root"
-import { IRequirementTemplateSectionAttributes } from "../../../../types/api-request"
-import { EditableInputWithControls } from "../../../shared/editable-input-with-controls"
-import { RemoveConfirmationModal } from "../../../shared/remove-confirmation-modal"
-import { RequirementBlockAccordion } from "../../requirements-library/requirement-block-accordion"
-import { RequirementsLibraryDrawer } from "../../requirements-library/requirements-library-drawer"
+import { useMst } from "../../../../../setup/root"
+import { IRequirementTemplateSectionAttributes } from "../../../../../types/api-request"
+import { EditableInputWithControls } from "../../../../shared/editable-input-with-controls"
+import { RemoveConfirmationModal } from "../../../../shared/remove-confirmation-modal"
+import { RequirementBlockAccordion } from "../../../requirements-library/requirement-block-accordion"
+import { RequirementsLibraryDrawer } from "../../../requirements-library/requirements-library-drawer"
 import { IRequirementTemplateForm, formScrollToId } from "./index"
 
 interface IProps {
