+++ app/frontend/components/domains/requirement-template/screens/edit-requirement-template-screen/sections-dnd/droppable-section.tsx
