+++ app/frontend/components/domains/requirement-template/screens/edit-requirement-template-screen/sections-dnd/multipleContainersKeyboardCoordinates.ts
