+++ app/frontend/components/domains/requirement-template/screens/edit-requirement-template-screen/sections-dnd/sortable-requirement-block.tsx
