+++ app/frontend/components/domains/requirement-template/screens/jurisdiction-digital-permit-screen/digital-building-permits-list.tsx
@@ -3,11 +3,11 @@ import { format } from "date-fns"
 import { observer } from "mobx-react-lite"
 import React from "react"
 import { useTranslation } from "react-i18next"
-import { useTemplateVersions } from "../../../../hooks/resources/use-template-versions"
-import { ErrorScreen } from "../../../shared/base/error-screen"
-import { LoadingScreen } from "../../../shared/base/loading-screen"
-import { RouterLink } from "../../../shared/navigation/router-link"
-import { VersionTag } from "../../../shared/version-tag"
+import { useTemplateVersions } from "../../../../../hooks/resources/use-template-versions"
+import { ErrorScreen } from "../../../../shared/base/error-screen"
+import { LoadingScreen } from "../../../../shared/base/loading-screen"
+import { RouterLink } from "../../../../shared/navigation/router-link"
+import { VersionTag } from "../../../../shared/version-tag"
 
 interface IProps {
   activityId?: string
