+++ app/frontend/components/domains/requirement-template/screens/jurisdiction-digital-permit-screen/index.tsx
@@ -3,10 +3,10 @@ import { observer } from "mobx-react-lite"
 import React, { useEffect } from "react"
 import { useTranslation } from "react-i18next"
 import { useSearchParams } from "react-router-dom"
-import { useActivityOptions } from "../../../../hooks/resources/use-activity-options"
-import { useMst } from "../../../../setup/root"
-import { ErrorScreen } from "../../../shared/base/error-screen"
-import { LoadingScreen } from "../../../shared/base/loading-screen"
+import { useActivityOptions } from "../../../../../hooks/resources/use-activity-options"
+import { useMst } from "../../../../../setup/root"
+import { ErrorScreen } from "../../../../shared/base/error-screen"
+import { LoadingScreen } from "../../../../shared/base/loading-screen"
 import { DigitalBuildingPermitsList } from "./digital-building-permits-list"
 
 const sharedTabTextStyles = {
