+++ app/frontend/components/domains/requirement-template/screens/requirement-template-screen.tsx
@@ -2,18 +2,18 @@ import { Box, Center, Container, Flex, Heading, Text, VStack } from "@chakra-ui/
 import { observer } from "mobx-react-lite"
 import React from "react"
 import { useTranslation } from "react-i18next"
-import { useSearch } from "../../../hooks/use-search"
-import { useMst } from "../../../setup/root"
-import { Paginator } from "../../shared/base/inputs/paginator"
-import { PerPageSelect } from "../../shared/base/inputs/per-page-select"
-import { SharedSpinner } from "../../shared/base/shared-spinner"
-import { ToggleArchivedButton } from "../../shared/buttons/show-archived-button"
-import { SearchGrid } from "../../shared/grid/search-grid"
-import { SearchGridItem } from "../../shared/grid/search-grid-item"
-import { RouterLinkButton } from "../../shared/navigation/router-link-button"
-import { VersionTag } from "../../shared/version-tag"
-import { GridHeaders } from "./grid-header"
-import { TenmplateVersionsSidebar } from "./template-versions-sidebar"
+import { useSearch } from "../../../../hooks/use-search"
+import { useMst } from "../../../../setup/root"
+import { Paginator } from "../../../shared/base/inputs/paginator"
+import { PerPageSelect } from "../../../shared/base/inputs/per-page-select"
+import { SharedSpinner } from "../../../shared/base/shared-spinner"
+import { ToggleArchivedButton } from "../../../shared/buttons/show-archived-button"
+import { SearchGrid } from "../../../shared/grid/search-grid"
+import { SearchGridItem } from "../../../shared/grid/search-grid-item"
+import { RouterLinkButton } from "../../../shared/navigation/router-link-button"
+import { VersionTag } from "../../../shared/version-tag"
+import { GridHeaders } from "../grid-header"
+import { TenmplateVersionsSidebar } from "../template-versions-sidebar"
 
 export const RequirementTemplatesScreen = observer(function RequirementTemplate() {
   const { requirementTemplateStore } = useMst()
