+++ app/frontend/components/domains/requirement-template/screens/template-version-screen.tsx
@@ -8,10 +8,10 @@ import { useTemplateVersion } from "../../../../hooks/resources/use-template-ver
 import { ErrorScreen } from "../../../shared/base/error-screen"
 import { LoadingScreen } from "../../../shared/base/loading-screen"
 import { BuilderFloatingButtons } from "../builder-floating-buttons"
-import { BuilderHeader } from "../edit-requirement-template-screen/builder-header"
+import { SectionsDisplay } from "../sections-display"
 import { SectionsSidebar } from "../sections-sidebar"
 import { useSectionHighlight } from "../use-section-highlight"
-import { SectionsDisplay } from "./sections-display"
+import { BuilderHeader } from "./edit-requirement-template-screen/builder-header"
 
 const scrollToIdPrefix = "template-version-scroll-to-id-"
 export const formScrollToId = (id: string) => `${scrollToIdPrefix}${id}`
