+++ app/frontend/components/domains/requirement-template/sections-display.tsx
@@ -1,9 +1,9 @@
 import { Box, HStack, Stack, Text } from "@chakra-ui/react"
 import { observer } from "mobx-react-lite"
 import React from "react"
-import { IDenormalizedRequirementTemplateSection } from "../../../../types/types"
-import { RequirementBlockAccordion } from "../../requirements-library/requirement-block-accordion"
-import { formScrollToId } from "./template-version-screen"
+import { IDenormalizedRequirementTemplateSection } from "../../../types/types"
+import { RequirementBlockAccordion } from "../requirements-library/requirement-block-accordion"
+import { formScrollToId } from "./screens/template-version-screen"
 
 interface IProps {
   shouldCollapseAll?: boolean
